feat: split mongo storage from init
